refactor: 规范属性折叠组件名称
