+++ src/components/LayerAttribute/BubbleLayerStyle/index.tsx
@@ -1,15 +1,10 @@
 import classNames from 'classnames';
 import React, { memo, useMemo } from 'react';
-import { Form, FormCollapse, FormItem, Input, NumberPicker, Select, Switch } from '@formily/antd';
+import { Form, FormItem, Input, NumberPicker, Select, Switch } from '@formily/antd';
 import { createSchemaField } from '@formily/react';
 import type { Form as FormInstance } from '@formily/core';
 import { createForm, onFormValuesChange } from '@formily/core';
-import Collapse from '../components/Collapse';
-import FieldSelect from '../components/FieldSelect';
-import ColorPicker from '../components/ColorPicker';
-import RibbonSelect from '../components/RibbonSelect';
-import Slider from '../components/Slider';
-import SliderRange from '../components/SliderRange';
+import { FormCollapse, FieldSelect, ColorPicker, RibbonSelect, Slider, SliderRange } from '../components';
 import type { BubbleLayerStyleAttributeProps } from './types';
 import schema from './schema';
 import { bubbleLayerStyleConfigToFlat, bubbleLayerStyleFlatToConfig } from './helper';
@@ -31,7 +26,6 @@ export const BubbleLayerStyleAttributeSchemaField: React.FC<Pick<BubbleLayerStyl
           Slider,
           RibbonSelect,
           ColorPicker,
-          Collapse,
           FieldSelect,
           SliderRange,
         },
