+++ src/components/LayerAttribute/components/FormCollapse/index.less
@@ -1,7 +1,7 @@
 @root-entry-name: 'default';
 @import (reference) '~antd/es/style/themes/index.less';
 
-@collapse-prefix-cls: ~'@{ant-prefix}-custom-formily-collapse';
+@collapse-prefix-cls: ~'@{ant-prefix}-custom-formily-form-collapse';
 
 .@{collapse-prefix-cls} {
   .@{ant-prefix}-collapse-item {
